requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.9
configparser==7.1.0
Django==5.1
django-environ==0.11.2
djangorestframework==3.15.2
docx==0.2.4
etelemetry==0.3.1
filelock==3.16.1
fitz==0.0.1.dev2
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.62.3
gunicorn==23.0.0
httplib2==0.22.0
idna==3.8
isodate==0.6.1
looseversion==1.3.0
lxml==5.3.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
numpy==2.1.1
packaging==24.1
pandas==2.2.3
pathlib==1.0.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==4.25.4
prov==2.0.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pydot==3.0.1
pyparsing==3.1.4
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
pyxnat==1.6.2
rdflib==6.3.2
requests==2.32.3
rsa==4.9
scipy==1.14.1
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.1
tqdm==4.66.5
traits==6.3.2
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
